requirements.txt
streamlit
fastapi
uvicorn
langchain
langchain-community
pypdf
chromadb
python-dotenv
langchain-openai
